component: 各コンポーネントを分割し、components内に移動 #5
