+++ src/package.json
@@ -22,9 +22,12 @@
     "clsx": "^2.1.1",
     "jotai": "^2.12.4",
     "lucide-react": "^0.509.0",
+    "mermaid": "^11.6.0",
     "next": "15.3.2",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-markdown": "^10.1.0",
+    "rehype-sanitize": "^6.0.0",
     "tailwind-merge": "^3.2.0",
     "three": "^0.176.0"
   },
