+++ src/src/app/(main)/layout.tsx
@@ -1,68 +1,11 @@
-import { InfoIcon, PackagePlusIcon } from "lucide-react";
-import Link from "next/link";
 import React from "react";
 
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
-import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
+import { Header } from "@/components/layout/header";
 
 const MainLayout = ({ children }: Readonly<{ children: React.ReactNode }>) => {
   return (
     <>
-      <header>
-        <div className="h-16 px-4 md:px-10 flex justify-between items-center font-geist-sans">
-          <Dialog>
-            <DialogTrigger asChild>
-              <Button variant={"ghost"} size={"icon"} className="ml-auto hidden ">
-                <PackagePlusIcon />
-              </Button>
-            </DialogTrigger>
-            <DialogContent className="font-geist-sans">
-              <DialogHeader>
-                <DialogTitle>プロジェクトの追加</DialogTitle>
-                <DialogDescription>必要な情報を記入してください</DialogDescription>
-              </DialogHeader>
-              <Button className="bg-emerald-500 hover:bg-emerald-600">プロジェクトを追加</Button>
-            </DialogContent>
-          </Dialog>
-          <Button variant={"ghost"} size={"icon"} className="hidden">
-            <InfoIcon />
-          </Button>
-          <DropdownMenu>
-            <DropdownMenuTrigger className="hidden" asChild>
-              <Avatar>
-                <AvatarImage src="https://github.com/shadcn.png" alt="@shadcn" />
-                <AvatarFallback>CN</AvatarFallback>
-              </Avatar>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent align="end" className="min-w-64 max-w-64 font-geist-sans">
-              <DropdownMenuLabel className="py-4">My Account</DropdownMenuLabel>
-              <DropdownMenuSeparator />
-              <DropdownMenuItem>Githubアカウントページ</DropdownMenuItem>
-              <DropdownMenuItem>ログアウト</DropdownMenuItem>
-            </DropdownMenuContent>
-          </DropdownMenu>
-          <h1 className="font-semibold text-2xl font-cardo cursor-default">
-            <Link href="/">RIPro</Link>
-          </h1>
-          <Button className="bg-emerald-500 hover:bg-emerald-600">ログイン</Button>
-        </div>
-      </header>
+      <Header />
       <main className="w-full h-[calc(100%-64px)] font-geist-sans">{children}</main>
     </>
   );
