+++ src/src/config/cemetery.ts
@@ -76,3 +76,15 @@ export const CEMETERY_POSITIONS: { position: Vector3 }[] = [
   { position: [30, 0.25, 42] },
   { position: [40, 0.25, 42] },
 ];
+
+export const CEMETERY_PROJECTS = CEMETERY_POSITIONS.map(({position}, i) => ({
+  title: "リポジトリ" + (i + 1),
+  descripton: "説明" + (i + 1),
+  tags: [
+    {
+      label: "未完成"
+    }
+  ],
+  reflection: "反省文" + (i + 1),
+  position: position
+}))
