+++ src/src/config/road.ts
@@ -1,0 +1,52 @@
+import { Vector3 } from "@react-three/fiber";
+
+export const ROAD_POSITIONS: { position: Vector3 }[] = [
+  { position: [1.5, -0.5, -40] },
+  { position: [1.5, -0.5, -36] },
+  { position: [1.5, -0.5, -32] },
+  { position: [1.5, -0.5, -28] },
+  { position: [1.5, -0.5, -24] },
+  { position: [1.5, -0.5, -20] },
+  { position: [1.5, -0.5, -16] },
+  { position: [1.5, -0.5, -12] },
+  { position: [1.5, -0.5, -8] },
+  { position: [1.5, -0.5, -4] },
+  { position: [1.5, -0.5, 0] },
+  { position: [1.5, -0.5, 4] },
+  { position: [1.5, -0.5, 8] },
+  { position: [1.5, -0.5, 12] },
+  { position: [1.5, -0.5, 16] },
+  { position: [1.5, -0.5, 20] },
+  { position: [1.5, -0.5, 24] },
+  { position: [1.5, -0.5, 28] },
+  { position: [1.5, -0.5, 32] },
+  { position: [1.5, -0.5, 36] },
+  { position: [1.5, -0.5, 40] },
+  { position: [1.5, -0.5, 44] },
+  { position: [1.5, -0.5, 48] },
+
+  //
+  { position: [-1.5, -0.5, -40] },
+  { position: [-1.5, -0.5, -36] },
+  { position: [-1.5, -0.5, -32] },
+  { position: [-1.5, -0.5, -28] },
+  { position: [-1.5, -0.5, -24] },
+  { position: [-1.5, -0.5, -20] },
+  { position: [-1.5, -0.5, -16] },
+  { position: [-1.5, -0.5, -12] },
+  { position: [-1.5, -0.5, -8] },
+  { position: [-1.5, -0.5, -4] },
+  { position: [-1.5, -0.5, 0] },
+  { position: [-1.5, -0.5, 4] },
+  { position: [-1.5, -0.5, 8] },
+  { position: [-1.5, -0.5, 12] },
+  { position: [-1.5, -0.5, 16] },
+  { position: [-1.5, -0.5, 20] },
+  { position: [-1.5, -0.5, 24] },
+  { position: [-1.5, -0.5, 28] },
+  { position: [-1.5, -0.5, 32] },
+  { position: [-1.5, -0.5, 36] },
+  { position: [-1.5, -0.5, 40] },
+  { position: [-1.5, -0.5, 44] },
+  { position: [-1.5, -0.5, 48] },
+];
